+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The file contains important changes from version to version
 
+
+## v0.1.0 - (2023.04.07 - 2024.04.24)
+
+### Features
+
+* Added new domain zone: `any`
+* File logging (scrapy + selenium)
+* Added looking for any matched query on the page
+* Added processing js files
+* Added support of multiple built-in web-browsers (Chrome/Chromium, Firefox, Edge, Safari)
+
 ## v0.1.0 - (2023.03.13 - 2023.04.07)
 
 ### Features